Remove "close_old_db_connections" logic, as unittests were failing with it
This decorator was only implemented to try to troubleshoot Django
Channels issues we were having. However, Django 3.0 provided error
output that lead to fixing a few channels releated errors.

As result, it's possible we don't need this decorator anymore.
Going to test with the decorator commented out. If we have no errors
for a after this, then it's safe to remove the decorator entirely.