Revert custom logging file handler for now
It works fine for long running processes.
However, it turns out that it breaks horribly for any python
programs that stop and start repeatedly.

It seems that every time a program closes and starts back up, the
handler tends to make a new file, which is not what we want.
Needs further debugging.