Document fix for response.POST submitting strangely for dynamic form input
Technically it doesn't seem to affect this specific form example
in CaeHome. However, it does affect other forms, such as in
GradApps. This fix should make the form work in both instances,
so it just unconditionally behaves as expected.

Documenting it in workspace also makes it more visibly available
to other projects.